+++ CHANGELOG.rst
@@ -3,12 +3,15 @@ Changelog
 
 Version 2.2.0
 -------------
+- Don't force loading of neurons into memory for Population (#922). See new API of
+  :class:`Population<neurom.core.population.Population>` and
+  :func:`load_neurons<neurom.io.utils.load_neurons>`
 - Move ``total_length`` feature to from ``neuritefunc`` to ``neuronfunc``. Use ``neurite_lengths``
   feature for neurites
 - Include morphology filename extension into Neuron's name
 - Extend ``tree_type_checker`` to accept a single tuple as an argument. Additionally validate
-  function's arguments (#909, #912)
-- Optimize Sholl analysis code (#905)
+  function's arguments (#912, #914)
+- Optimize Sholl analysis code (#905, #919)
 
 Version 2.1.2
 -------------